♻️ refactor : 배경 애니메이션 및 파티클 시스템을 위한 컴포넌트 분리 #1
